backend/requirements.txt
Django>=4.2
djangorestframework
django-cors-headers
gunicorn
dj-database-url
psycopg2-binary
whitenoise
